breakdown the code
